refactor: move client code to dedicated package
